% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{countdata}
\alias{countdata}
\title{Example RNA-Seq count data.}
\format{A data frame with 18094 rows and 17 variables:
\describe{
 \item{Geneid}{ensembl gene ID}
 \item{MGIID}{MGI ID}
 \item{MGIsymbol}{MGIsymbol}
 }}
\usage{
countdata
}
\description{
A dataset containing the RNA-Seq counts, a typical output of
RNA-Seq bioinformatics algorithms. Each row is a gene identifier and each
sample is represented by a column. Some gene information variables (various
IDs, chromosome, start and end positions, etc.) are provided. There are three
tissues (A,B,C) each with three biological replicates.
}
\keyword{datasets}

