% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_count_data}
\alias{get_count_data}
\title{Retrieve the counts from a data set.}
\usage{
get_count_data(groupvec, data, geneid_ind = 1)
}
\description{
This function retrieves a data set containing the vectors of counts and
one gene identifier.
}
\examples{
mycounts <- get_count_data(groupvec=c("tissueA","tissueB","tissueC"),data=countdata, geneid_ind=1)
head(mycounts)
}

