% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_count_data_long}
\alias{get_count_data_long}
\title{Creates long format count data.}
\usage{
get_count_data_long(groupvec, id_vars, data)
}
\arguments{
\item{groupvec}{a character vector of group names (i.e. tissue type), assuming
column names of counts include these characters}

\item{id_vars}{a character vector of gene identifier column names}

\item{data}{a data frame that includes counts and gene identifiers}
}
\description{
Creates long format count data.
}
\examples{
mycountslong <- get_count_data_long(groupvec=c("tissueA","tissueB","tissueC"),data=countdata, id_vars = c("Geneid","MGIID","Chr"))
head(mycountslong)
}

